Fixed bug for order numbers with non-numerical prefix. Cast IDs to int for improved security. Updated Mastercard 3d image.

+++ complete.php
@@ -6,7 +6,7 @@
  *  @copyright 2015 Quickpay
  *  @license   http://opensource.org/licenses/afl-3.0.php  Academic Free License (AFL 3.0)
  *
- *  $Date: 2016/11/12 20:35:47 $
+ *  $Date: 2016/12/24 07:34:08 $
  *  E-mail: helpdesk@quickpay.net
  */
 
@@ -18,8 +18,8 @@ if (_PS_VERSION_ >= '1.5.0.0') {
     die('Bad version');
 }
 
-$id_cart = Tools::getValue('id_cart');
-$id_module = Tools::getValue('id_module');
+$id_cart = (int)Tools::getValue('id_cart');
+$id_module = (int)Tools::getValue('id_module');
 $key = Tools::getValue('key');
 if (!$id_module || !$key) {
     Tools::redirect('history.php');
